requirements.txt
pygame-ce>=2.4.0
opencv-python>=4.8.0
